test: add property-based fuzzing for HIP parsers (fixes 3 build-crash bugs)
Adds fast-check property-based tests for the HIP frontmatter/markdown parsing
helpers, which process untrusted content from open draft-HIP PRs. Satisfies the
OpenSSF Scorecard Fuzzing check (property-based testing in JavaScript) and, more
importantly, hardens the build against malformed PR input.

To make the parsers testable in isolation, the pure helpers (parseMarkdown,
extractHip, parsePositiveInteger, draftHipNumber, replaceHipImages) move from
build-data.js into a new side-effect-free module site/scripts/hip-parse.js;
build-data.js imports them, with no change to build behavior.

The new tests immediately found three inputs that crash the site build:
- replaceHipImages("![x](../assets/%)")  -> URIError from decodeURIComponent
- parsePositiveInteger({ toString: [] }) -> TypeError from String()
- draftHipNumber(...)                     -> same, via parsePositiveInteger

All three now fall back gracefully instead of throwing, so a single malformed
draft HIP can no longer abort the whole build. Run with `npm test` in site/.
